TNO-2096 Extract formatDate function
