+++ app/subscriber/src/features/search-page/SearchPage.tsx
@@ -25,20 +25,7 @@ import {
 import { AdvancedSearch } from './components';
 import { Player } from './player/Player';
 import * as styled from './styled';
-import { filterFormat } from './utils';
-
-const formatDate = (date: string) => {
-  const d = new Date(date);
-  const day = d.toLocaleDateString('en-us', { day: '2-digit' });
-  const month = d.toLocaleDateString('en-us', { month: 'short' }).toUpperCase();
-  const year = d.toLocaleDateString('en-us', { year: 'numeric' });
-  const hour = new Intl.DateTimeFormat('en-us', {
-    hour12: false,
-    hour: '2-digit',
-    minute: 'numeric',
-  }).format(d);
-  return `${day}-${month}-${year}, ${hour}`;
-};
+import { filterFormat, formatDate } from './utils';
 
 export interface ISearchType {
   showAdvanced?: boolean;
