+++ app/subscriber/src/features/search-page/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './filterFormat';
+export * from './formatDate';
 export * from './formatSearch';
 export * from './stripHtml';
 export * from './trimWords';
